Improves the process of finding the correct gps location of the video. I have reviewed all my gopro videos and made all improvements to improve accuracy I have removed the GpsFix and because some videos recorded with the gopro have the correct location but the gpsfix flag was not active, the same thing also happened with the speed that there were videos with the correct location but at speed 0
Then I have added a process to select the most repeated and next locations so we discard false locations.
we also control the height because it is an indicator of incorrect position.

That's why in the end I added a filter by country, so that they discard the false locations, but this generates an infinite loop that makes the videos repeat over and over again.
I have gone crazy and I have not been able to solve it

Please help

All the best