switch to retryablehttp to allow downloading many files
My HERO12 did not like being asked for more than about 5-7 files
simultaneously, but dropped some of the /gp/gpMediaMetadata with a
timeout. So after the files where the call to gpMediaMetadata was
successfull were downloaded, I got presented with a bunch of errors
where it didn't work, and of cause the files were not downloaded.

Switching to a retrying approach means that we still hammer the GoPro
for the first couple seconds, but allow the requests to proceed later.

Some of this is still WIP - the config file has a timeout option, that I
dropped since I do not understand enough go to incorporate it, yet.